+++ gradle.properties
@@ -1,4 +1,4 @@
-mod_version=2.3.2
+mod_version=1.6.0
 mod_name=classyhats
 mod_fancy_name=ClassyHats
 mod_author=wiresegal
@@ -7,9 +7,8 @@ mod_author=wiresegal
 mc_version=1.12.2
 forge_version=14.23.5.2768
 mcp_version=stable_39
-
-kt_version=1.2.0
-liblib_version=4.18
+kt_version=1.3.0
+liblib_version=4.19
 jei_version=4.9.1.175
 
 org.gradle.jvmargs=-Xmx3G
